Refactor `setup()` to avoid race condition
The render loop starts immediately after the `setup()` method returns.
If `setup()` is async, it returns immediately after the first `await`.
Due to the event-based nature of the initialition and render loop, the
first call to `draw()` sometimes happens before the microphone is
initialized which causes the app to crash.

This patch fixes the issue by factoring out the asynchronous part and
ignoring the not yet initialized elements in the draw call.