Replace custom sphere geometry generation with three.js built-in
Since the deformation of the geometry is now computed on the GPU with
analytic normals, the custom indexed face-topology preserving geometry
generation is no longer needed (even though rendering the custom
geometry would be slightly more efficient since it does not include
duplicate vertices).

The `detail` parameter has been changed in the process from an
exponential towards a more linear behavior with respect to the number of
faces.